mmc: host: cmdq: Proceed with other interrupts if CQTCN is zero
There are instances observed that CQ triggers interrupt for halt
operation with both halt bit and task-complete bits set in CQIS
register. Since this task-complete interrupt is a spurious one,
TQTCN remains zero. In this case, we skip servicing halt interrupt
and context which is trying to halt CQ gets timeout while waiting
for halt interrupt.

To cover this scenario, proceed with servicing halt interrupt even
when it gets the spurious task-complete interrupt.

Change-Id: Ibe4e3b0a1a931ec6eea3f23ae239da50f462bc0b
Signed-off-by: Veerabhadrarao Badiganti <vbadigan@codeaurora.org>