usb: pd: Allow sink operations to queue when not SinkTxOk
Currently user requested operations such as select_pdo,
get_src_cap_ext, DR/PR swap when done in sink mode will get
rejected simply because the current Rp state is SinkTxNG. But
since these APIs' semantics effectively queue the request to
be handled in usbpd_sm(), and there is already a check within
PE_SNK_Ready to send only during SinkTxOK, the requests don't
need to be rejected and should still be allowed. Since there
is a blocking wait, if the Source relinquishes Rp and allows
for the AMS, then the operation should be able to go through.

Change-Id: I79e78d3f28fdbb3d8e43017181981a78d3decf84
Signed-off-by: Jack Pham <jackp@codeaurora.org>